% Generated by roxygen2 (4.0.0): do not edit by hand
\name{funcThresh}
\alias{funcThresh}
\title{Function to calculate correlation threshold in parallel.}
\usage{
funcThresh(.x, Y, Weights, Factor, anno, index.ref, thresholds, set.size)
}
\arguments{
\item{.x}{Vector.}

\item{Y}{Matrix.}

\item{Weights}{A object of class \code{Weights} or a list of weights.}

\item{Factor}{Character string.}

\item{anno}{Dataframe.}

\item{index.ref}{Vector.}

\item{thresholds}{Vector.}

\item{set.size}{Integer.}
}
\value{
Matrix.
}
\description{
Internal function for parallel computing.
}
\author{
Saskia Freytag
}
\keyword{internal}

