% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.optimizeParameters}
\alias{is.optimizeParameters}
\title{Checking \code{optimizeParameters} class.}
\usage{
is.optimizeParameters(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is.optimizeParameters} returns a logical scalar;
\code{TRUE} if the object is of the class \code{optimizeParameters}.
}
\description{
\code{is.optimizeParameters} checks if object is of \code{optimizeParameters} class.
}
\examples{
\donttest{
Y<-simulateGEdata(500, 500, 10, 2, 5, g=NULL, Sigma.eps=0.1,
250, 100, intercept=FALSE, check.input=FALSE)
opt<-optimizeParameters(Y, kW.hat=c(1,5,10), nu.hat=c(100,1000),
nc_index=251:500, methods=c("fnorm"), cpus=1, parallel=FALSE)
opt
is.optimizeParameters(opt)
}
}
\author{
Saskia Freytag
}
\seealso{
\code{\link{optimizeParameters}}
}

