% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.simulateGEdata}
\alias{is.simulateGEdata}
\title{Checking \code{simulateGEdata} class.}
\usage{
is.simulateGEdata(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is.simulateGEdata} returns a logical scaler;
\code{TRUE} if the object is of the class \code{simulateGEdata}.
}
\description{
\code{is.simulateGEdata} checks if object is of \code{simulateGEdata} class.
}
\examples{
\donttest{
Y<-simulateGEdata(500, 500, 10, 2, 5, g=NULL, Sigma.eps=0.1,
250, 100, intercept=TRUE, check.input=TRUE)
is.simulateGEdata(Y)
}
}
\author{
Saskia Freytag
}
\seealso{
\code{\link{simulateGEdata}}
}

