% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.optimizeParameters.R
\name{is.Threshold}
\alias{is.Threshold}
\title{Checking \code{Threshold} class.}
\usage{
is.Threshold(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is.Threshold} returns a logical scalar; 
\code{TRUE} if the object is of the class \code{Threshold}.
}
\description{
\code{is.Threshold} checks if object is of \code{Threshold} class.
}
\seealso{
\code{\link{calculateThreshold}}
}
\author{
Saskia Freytag
}
