% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.simulateGEdata.R
\name{is.Weights}
\alias{is.Weights}
\title{Checking \code{Weights} class.}
\usage{
is.Weights(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{is.Weights} returns a logical scaler; 
\code{TRUE} if the object is of the class \code{Weights}.
}
\description{
\code{is.Weights} checks if object is of \code{Weights} class.
}
\seealso{
\code{\link{findWeights}}
}
\author{
Saskia Freytag
}
