% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makePosSemiDef}
\alias{makePosSemiDef}
\title{Makes square matrices positive semi-definite.}
\usage{
makePosSemiDef(a, offset = 0)
}
\arguments{
\item{a}{Square matrix.}

\item{offset}{Offset.}
}
\value{
Positive semi-definite matrix.
}
\description{
Internal function which returns closest positive semi-definite matrix to input matrix.
}
\author{
Saskia Freytag
}
\keyword{internal}

