% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mashUp.R
\name{mashUp}
\alias{mashUp}
\title{Joining two correlation matrices by diagonal.}
\usage{
mashUp(true, est, plot.genes)
}
\arguments{
\item{true}{Matrix.}

\item{est}{Matrix.}

\item{plot.genes}{Vector of indices.}
}
\value{
Matrix.
}
\description{
Internal function that joins two matrices at their diagonal.
}
\author{
Saskia Freytag
}
\keyword{internal}

