% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotEFDR.R
\name{plotEFDR}
\alias{plotEFDR}
\title{Plots an object of class \code{EFDR}.}
\usage{
plotEFDR(x, main = "", legend, col = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{EFDR} or a list of objects of class \code{EFDR}.}

\item{main}{A character string describing the title of the plot.}

\item{legend}{A vector of character strings decribing the different EDFR objects in
\code{x}; only applicable when x is a list.}

\item{col}{A vector giving the colors, if \code{NULL} colors
are generated automatically.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{plot.EDFR} returns a plot.
}
\description{
\code{plotEFDR} plots the EDFR curves stored in the objects of class \code{EDFR}.
}
\examples{
Y<-simulateGEdata(500, 500, 10, 2, 5, g=NULL, Sigma.eps=0.1,
250, 100, intercept=FALSE, check.input=FALSE)
anno<-as.matrix(sample(1:4, dim(Y$Y)[1], replace=TRUE))
colnames(anno)<-"Factor"
weights<-findWeights(Y$Y, anno, "Factor")
efdr<-calculateEFDR(Y$Y, exclude=1:100, index.ref=1:10,
Weights=weights, anno=anno, Factor="Factor")
plotEFDR(efdr)
}
\author{
Saskia Freytag
}

