% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.simulateGEdata.R
\name{print.simulateGEdata}
\alias{print.simulateGEdata}
\title{Print an object of class \code{simualteGEdata}.}
\usage{
\method{print}{simulateGEdata}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{simulateGEdata}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{print.simualteGEdata} returns the information about simulation and
the first 5 rows and 5 columns of all matrices.
}
\description{
\code{print.simualteGEdata} is the \code{print} generic
for object so f the class \code{simulateGEdata}.
}
\examples{
\donttest{
Y<-simulateGEdata(500, 500, 10, 2, 5, g=NULL, Sigma.eps=0.1,
250, 100, intrecept=TRUE, check.input=FALSE)
Y
}
}
\author{
Saskia Freytag
}
\seealso{
\code{\link{simulateGEdata}}
}

