% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUVcorr.R
\docType{package}
\name{RUVcorr}
\alias{RUVcorr}
\title{Removal of unwanted variation for gene-gene correlations.}
\description{
\pkg{RUVcorr} allows to apply global removal of 
unwanted variation (ridged version of RUV) to real and simulated gene expression data.
}
\details{
All gene expression data are assumed to be in the following format:
  \itemize{
    \item Rows correspond to arrays. 
    \item Columns correspond to genes.
  }
}
\author{
Saskia Freytag
}
