% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splitByFactor.R
\name{splitByFactor}
\alias{splitByFactor}
\title{Splitting a data set by a particular factor.}
\usage{
splitByFactor(X, anno, Factor)
}
\arguments{
\item{X}{A matrix containing gene expressions.}

\item{anno}{A dataframe or a matrix containing the annotation of arrays in \code{X}.}

\item{Factor}{A character string corresponding to a column name of \code{anno} to be used for splitting.}
}
\value{
\code{splitByFactor} returns a list object.
}
\description{
Internal function that splits a data set according to a particular factor.
}
\author{
Saskia Freytag
}
\keyword{internal}

