% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findIQR.R
\name{findIQR}
\alias{findIQR}
\title{Find the inter quantile range.}
\usage{
findIQR(x)
}
\arguments{
\item{Vector}{of gene expression values.}
}
\value{
Numeric value.
}
\description{
Internal function to find the inter quantile range.
}
\author{
Saskia Freytag
}
\keyword{internal}

