% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findMinmaxSamples.R
\name{findMinmaxSamples}
\alias{findMinmaxSamples}
\title{Find minimum and maximum samples in gene expression data.}
\usage{
findMinmaxSamples(x)
}
\arguments{
\item{x}{Matrix of gene expression values.}
}
\value{
Vector of indices.
}
\description{
Internal function that returns 5 samples with smallest inter-quantile range
and 5 samples with highest inter-quantile range.
}
\author{
Saskia Freytag
}
\keyword{internal}

