% Generated by roxygen2 (4.0.0): do not edit by hand
\name{funcPara}
\alias{funcPara}
\title{Function to optimize parameters in parallel.}
\usage{
funcPara(x, Y, nc_index, center = TRUE, index, methods)
}
\arguments{
\item{x}{Vector.}

\item{Y}{\code{simulateGE} object.}

\item{nc_index}{Vector.}

\item{center}{Logical.}

\item{index}{Vector.}

\item{methods}{Vector.}
}
\value{
List.
}
\description{
Internal function for parallel computing.
}
\author{
Saskia Freytag
}
\keyword{internal}

