% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeRankedList}
\alias{makeRankedList}
\title{Make ranked list of correlations.}
\usage{
makeRankedList(Data)
}
\arguments{
\item{Data}{matrix of gene-gene correlations.}
}
\value{
Matrix.
}
\description{
Internal function.
}
\author{
Saskia Freytag
}
\keyword{internal}

