% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{RUVcorr}
\alias{RUVcorr}
\alias{RUVcorr-package}
\title{Removal of unwanted variation for gene-gene correlations.}
\description{
\pkg{RUVcorr} allows to apply global removal of
unwanted variation (ridged version of RUV) to real and simulated gene expression data.
}
\details{
All gene expression data are assumed to be in the following format:
  \itemize{
    \item Rows correspond to arrays.
    \item Columns correspond to genes.
  }
}
\author{
Saskia Freytag
}

